+++ Python/example_controllers/vray_export/dynamic_objects.py
@@ -1,15 +1,13 @@
 from tdw.controller import Controller
 from tdw.add_ons.third_person_camera import ThirdPersonCamera
-from tdw.tdw_utils import TDWUtils
-from tdw.add_ons.vray_export import VRayExport
+from tdw.add_ons.vray_exporter import VRayExporter
 
 
-
-
-class PhotorealVRay(Controller):
+class DynamicObjects(Controller):
     """
     Create a typical TDW scene. Apply a force to the chair, then export using V-Ray add-on for maximum photorealism.
     """
+
     def __init__(self, port: int = 1071, check_version: bool = True, launch_build: bool = True):
         super().__init__(port=port, check_version=check_version, launch_build=launch_build)
         self.frame_range = 150
@@ -22,25 +20,18 @@ class PhotorealVRay(Controller):
                                    position={"x": 2.5, "y": 1.0, "z": 2},
                                    look_at={"x": 0, "y": 0.5, "z": 0.5},
                                    field_of_view=55)
-        export = VRayExport(image_width=1280, 
-                            image_height=720, 
-                            scene_name="tdw_room", 
-                            output_path="D:/VE2020_output/", 
-                            animate=True)
+        exporter = VRayExporter(image_width=1280,
+                                image_height=720,
+                                scene_name="tdw_room",
+                                output_directory="D:/VE2020_output/",
+                                animate=True)
         self.add_ons.append(camera)
-        # Set the resolution to 720p.
-        # Set render quality to maximum.
         # Load the scene.
         # Add the objects.
         # Adjust post-processing parameters to achieve a suitable depth of field and exposure.
         # Also adjust the ambient occlusion parameters for realistic shadowing in corners and under furniture objects.
         # Set shadow strength to near-full.
-        self.communicate([{"$type": "set_screen_size",
-                           "width": 1280,
-                           "height": 720},
-                          {"$type": "set_render_quality",
-                           "render_quality": 5},
-                          {"$type": "set_aperture",
+        self.communicate([{"$type": "set_aperture",
                            "aperture": 4.0},
                           {"$type": "set_focus_distance",
                            "focus_distance": 2.25},
@@ -58,41 +49,41 @@ class PhotorealVRay(Controller):
                                               position={"x":0.125, "y": 0, "z": 0.37},
                                               rotation={"x": 0, "y": 45, "z": 0}),
                          self.get_add_object(model_name="live_edge_coffee_table",
-                                              object_id=self.get_unique_id(),
-                                              position={"x": 1.81, "y": 0, "z": -0.47},
-                                              rotation={"x": 0, "y": -90, "z": 0}),
+                                             object_id=self.get_unique_id(),
+                                             position={"x": 1.81, "y": 0, "z": -0.47},
+                                             rotation={"x": 0, "y": -90, "z": 0}),
                          self.get_add_object(model_name="bastone_floor_lamp",
-                                              object_id=self.get_unique_id(),
-                                              position={"x": 2.35, "y": 0, "z": 1},
-                                              rotation={"x": 0, "y": 35, "z": 0}),
+                                             object_id=self.get_unique_id(),
+                                             position={"x": 2.35, "y": 0, "z": 1},
+                                             rotation={"x": 0, "y": 35, "z": 0}),
                          self.get_add_object(model_name="chair_eames_plastic_armchair",
-                                              object_id=self.chair_id,
-                                              position={"x": 0.1, "y": 0, "z": 1.85},
-                                              rotation={"x": 0, "y": 63.25, "z": 0}),
+                                             object_id=self.chair_id,
+                                             position={"x": 0.1, "y": 0, "z": 1.85},
+                                             rotation={"x": 0, "y": 63.25, "z": 0}),
                          self.get_add_object(model_name="vase_05",
-                                              object_id=self.get_unique_id(),
-                                              position={"x": 0.125, "y": 0.3960, "z": 0.37},
-                                              rotation={"x": 0, "y": 63.25, "z": 0}),
+                                             object_id=self.get_unique_id(),
+                                             position={"x": 0.125, "y": 0.3960, "z": 0.37},
+                                             rotation={"x": 0, "y": 63.25, "z": 0}),
                          self.get_add_object(model_name="zenblocks",
-                                              object_id=self.get_unique_id(),
-                                              position={"x": 1.8, "y": 0.303, "z": -0.517},
-                                              rotation={"x": 0, "y": 70, "z": 0})])
+                                             object_id=self.get_unique_id(),
+                                             position={"x": 1.8, "y": 0.303, "z": -0.517},
+                                             rotation={"x": 0, "y": 70, "z": 0})])
         # Let physics settle.       
         for step in range(15):
-            resp = self.communicate([])
-        self.add_ons.append(export)
+            self.communicate([])
+        self.add_ons.append(exporter)
         # Ten frame hold before moving chair.
         for step in range(10):
-            resp = self.communicate([])
+            self.communicate([])
         # Apply a force and run simulation for 150 frames.
         self.communicate({"$type": "apply_force_to_object",
-                           "id": self.chair_id,
-                           "force": {"x": 0, "y": 0.5, "z": -35}})
+                          "id": self.chair_id,
+                          "force": {"x": 0, "y": 0.5, "z": -35}})
         for step in range(self.frame_range):
-            resp = self.communicate([])
+            self.communicate([])
         # Launch Vantage render in headless mode; it will run to completion and automatically close.
-        export.launch_render()
+        exporter.launch_renderer()
         
 
 if __name__ == "__main__":
-    PhotorealVRay(launch_build=False).run()
+    DynamicObjects().run()
